# Improve performance of parser