--- a/.travis.yml
+++ b/.travis.yml
@@ -1,5 +1,7 @@
 language: go
-go: 
+go:
  - "1.11.x"
 env:
  - GO111MODULE=on
+script:
+  - script/cibuild