--- a/hesabu/parser.go
+++ b/hesabu/parser.go
@@ -4,7 +4,6 @@ import (
 	"fmt"
 	"log"
 	"math"
-	"regexp"
 	"strconv"
 	"strings"
 
@@ -107,10 +106,6 @@ func (parsedEquations ParsedEquations) newSingleError(key string, message string
 	return make(map[string]interface{}), &CustomError{EvalError: evalError}
 }
 
-var and_regex = regexp.MustCompile(`(?i)\bAND\b`)
-var or_regex = regexp.MustCompile(`(?i)\bOR\b`)
-var single_equals_regex = regexp.MustCompile(`(\b|\s)(=)(\b|\s)`)
-
 func isNumeric(s string) bool {
 	_, err := strconv.ParseFloat(s, 64)
 	return err == nil
@@ -123,19 +118,64 @@ func isNumeric(s string) bool {
 //       = => == (only for equality comparison)
 //
 func clean(expression string) (cleanExpression string) {
-	if isNumeric(expression) {
+	if !needsCleaning(expression) {
 		return expression
 	}
 	cleanExpression = expression
-	if strings.Contains(expression, "and") ||
-		strings.Contains(expression, "AND") ||
-		strings.Contains(expression, "OR") ||
-		strings.Contains(expression, "or") ||
-		strings.Contains(expression, "=") {
-		cleanExpression = and_regex.ReplaceAllString(cleanExpression, "&&")
-		cleanExpression = or_regex.ReplaceAllString(cleanExpression, "||")
-		cleanExpression = single_equals_regex.ReplaceAllString(cleanExpression, "==")
+	cleanExpression = strings.Replace(cleanExpression, " AND ", " && ", -1)
+	cleanExpression = strings.Replace(cleanExpression, " OR ", " || ", -1)
+	cleanExpression = strings.Replace(cleanExpression, " and ", " && ", -1)
+	cleanExpression = strings.Replace(cleanExpression, " or ", " || ", -1)
+
+
+	if strings.Contains(expression, "=") {
+		cleanExpression = replaceSingleEquals(expression)
 	}
 
 	return cleanExpression
+}
+
+func needsCleaning(expression string) bool {
+	// Numbers never need cleaning
+	if isNumeric(expression) {
+		return false
+	}
+
+	// Invalid formulas
+	if strings.HasPrefix(expression, "=") {
+		return false
+	}
+	if strings.HasSuffix(expression, "=") {
+		return false
+	}
+	return true
+}
+
+// Replaces a single '=' with its double cousing '==', it takes care
+// to not replace single equals that are actually a different symbol
+// ('==', '<=', '>=')
+//
+// Why don't you just use a regex? Turns out, regex in go is not that
+// fast. A regex replace for the single equals was on average almost a
+// third slower than not using a regex.
+func replaceSingleEquals(in string) string {
+	// Characters that combined with an equal form a special symbol
+	reserved := map[byte]int{'=': 1, '<': 1, '>': 1}
+	var t = []rune{}
+
+	// Loop over the runes from the in string
+	for pos, char := range in {
+		t = append(t, char)
+		if char == '=' {
+			// Make sure that neither the next, or the previous character is
+			// one of the reserveds
+			_, previousIsReserved := reserved[in[pos-1]]
+			_, nextIsReserved := reserved[in[pos+1]]
+			if !(previousIsReserved || nextIsReserved) {
+				t = append(t, '=')
+			}
+		}
+	}
+	// Convert runes back to string
+	return string(t)
 }