@mestachs pointed out that performance regressed by adding the cleaner in the parser.

This was mainly caused by the regexp that we were using, this PR replaces the regexes with a simple `Replace` (even though we call it multiple times, it's still fast) and a specialized function to replace single equals.

From the `hesabu`-gem specs:

```
// Before the clean function was added
Parsor and interpretor
  solve big problems
parsing 0.166265 seconds
solving 1.051677 seconds
parsing 0.166265 seconds, solving 1.051677 seconds
    parse and evaluate

Finished in 1.33 seconds (files took 0.26384 seconds to load)
```

```
// With the clean function added without regex
Parsor and interpretor
  solve big problems
parsing 0.189437 seconds
solving 1.043802 seconds
parsing 0.189437 seconds, solving 1.043802 seconds
    parse and evaluate

Finished in 1.34 seconds (files took 0.27855 seconds to load)
```

Or by running `go test -bench=. -run=^$ ./...` which will run the parser benchmark:

```
// Without Clean						: BenchmarkParse-4   	      10	 165449475 ns/op
// With replaceSingleEqual  : BenchmarkParse-4   	      10	 188343765 ns/op
// With regex								: BenchmarkParse-4   	       5	 318208619 ns/op
```

So the replaceSingleEqual is still 0.02s slower, but it is doing more, so I guess that's a win.

(This PR also brings in that `script/cibuild` will now be run by Travis, and a small test runner that runs each example file in the test folder and checks the output)